Add role be Authentication and Autherization

+++ Controller/Admin_control.py
@@ -1,4 +1,5 @@
 from Model import Database_Model
+from fastapi import HTTPException, status
 
 #Get all the user in the system
 
@@ -18,6 +19,10 @@ def add_new_movies(y,db):
 
 def update_movie(movie_name,request,db):
     update = db.query(Database_Model.Movie).filter(Database_Model.Movie.Title == movie_name).first()
+    if not update:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail= f"Could not found {movie_name}")
+
     update_film = request.model_dump(exclude_unset = True)
     for field, values in update_film.items():
         setattr(update, field, values)
@@ -28,6 +33,8 @@ def update_movie(movie_name,request,db):
 
 def delete_a_movie(id,db):
     movie_delete = db.query(Database_Model.Movie).filter(Database_Model.Movie.ID == id).first()
+    if not movie_delete:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Could not found movie id: {id}")
     db.delete(movie_delete)
     db.commit()
     return "Movie has been deleted"
